Hi, Ulelu Israel : 
    I really like this project of yours and I found while reading the source code that the performance of string contains in InvokeAsync can be optimized. Then remove the unused assignment code.

I created a set of test code to simulate the performance of the code before and after.
```cs
[GcForce(true)]
[MemoryDiagnoser]
[Orderer(SummaryOrderPolicy.FastestToSlowest)]
public class BenchInvoke
{
	private static readonly PathString Path = "/TestController/Other/";

	private static readonly string[] Blacklist = new[]
	{
		"Controller1",
		"Controller2",
		"Controller3",
		"Controller4",
		"Controller5"
	};
	
	[Benchmark(Baseline = true)]
	public bool Before()
	{
		if (!Path.ToString().Contains("WTCHDwatchpage") &&
		    !Path.ToString().Contains("watchdog") &&
		    !Path.ToString().Contains("WTCHDGstatics") &&
		    !Path.ToString().Contains("favicon") &&
		    !Path.ToString().Contains("wtchdlogger") &&
		    !Blacklist.Contains(Path.ToString().Remove(0, 1),
			    StringComparer.OrdinalIgnoreCase))
		{
			return true;
		}

		return false;
	}

	[Benchmark]
	public bool After()
	{
		var requestPath = Path.ToString().Remove(0,1);
		if (!requestPath.Contains("WTCHDwatchpage") &&
		    !requestPath.Contains("watchdog") &&
		    !requestPath.Contains("WTCHDGstatics") &&
		    !requestPath.Contains("favicon") &&
		    !requestPath.Contains("wtchdlogger") &&
		    !Blacklist.Contains(requestPath, StringComparer.OrdinalIgnoreCase))
		{
			return true;
		}

		return false;
	}
}
```
| Method |     Mean |   Error |  StdDev | Ratio |   Gen0 | Allocated | Alloc Ratio |
|------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
|  After | 217.6 ns | 1.81 ns | 1.52 ns |  0.27 | 0.0203 |      96 B |        1.00 |
| Before | 805.0 ns | 5.12 ns | 4.28 ns |  1.00 | 0.0200 |      96 B |        1.00 |

We can reuse the RequestPath and remove the '\\'directly，Thanks for reading！